refactor: rename nuxt-build to nuxt-pack
