+++ package.json
@@ -54,7 +54,7 @@
     "nuxt": "./bin/nuxt"
   },
   "scripts": {
-    "build": "node -r esm ./packages/nuxt-build/build.js",
+    "build": "node -r esm ./packages/nuxt-pack/build.js",
     "coverage": "codecov",
     "lint": "eslint --ext .js,.mjs,.vue bin/** benchmarks examples lib packages test",
     "postinstall": "opencollective || exit 0",
