+++ packages/nuxt-pack/build.js
