+++ packages/nuxt-pack/builtins.js
