+++ packages/nuxt-pack/package.js
