+++ rollup.config.js
@@ -1,4 +1,4 @@
-import config from './packages/nuxt-build/rollup.config'
+import config from './packages/nuxt-pack/rollup.config'
 
 export default config({
   rootDir: __dirname,
